Reuse existing voice channel categories
This allows us to remove the manage channels permission from the bot and avoid a scenario where a compromised bot deletes all the channels in the server.

Remove the strikes and voice channels tables, and the category class that aren't required anymore.
Remove the no longer required manage channels permission.