Reorganize buttons/modal into interactions
Having the code that triggers on button press sit in the code that defines the button was confusing. I don't want to go look at a button file to figure out how a modal is created. Or look at a modal file to figure out how a deployment is updated in the database.

Restructure the code around interactions where the interaction includes the entire interaction lifecycle, including the button press, the modal display and the submission event.

Only did new deployment and edit deployment in this commit.