Reset role select menu on errors and cooldown
Refactor role select menu interaction to use the updated flow with deferring the reply to show that the bot is thinking.

On error, reset the select menu to empty and indicate to the user that the interaction failed.
On success, clear the thinking indicator and update the deployment message with the new role assignment.

For cooldowns, also reset the select menu.

Remove reply in select menu interaction callback since we don't know what state the interaction is in and whether we can reply or not.