Force cached guild for all interactions
non cached interactions don't have a member. If/when it happens we will get a thrown exception that will be sent to the error log.
This seems to mostly happen when the bot restrarts mid interaction.
There is no way to re-fetch the whole interaction object and we would have to fetch the member and pipe it as a separate argument if we ever do want to handle these.